Re #787 Fixed MSVC 2005 warnings
git-svn-id: https://svn.code.sf.net/p/speed-dreams/code/trunk@5353 30fe4595-0a0c-4342-8851-515496e4dcbd

Former-commit-id: daabd11b0db72e3856ba0796079b374b873e097f
Former-commit-id: 2c9a3c68968f7dd72080ef80b0dd1bdc7e83fc3f

+++ src/libs/math/SGVec3.h
@@ -146,7 +146,7 @@ SGVec3<float>::fromGeod(const SGGeod& geod)
 {
   SGVec3<double> cart;
   SGGeodesy::SGGeodToCart(geod, cart);
-  return SGVec3<float>(cart(0), cart(1), cart(2));
+  return SGVec3<float>((float)cart(0), (float)cart(1), (float)cart(2));
 }
 
 template<>
@@ -166,7 +166,7 @@ SGVec3<float>::fromGeoc(const SGGeoc& geoc)
 {
   SGVec3<double> cart;
   SGGeodesy::SGGeocToCart(geoc, cart);
-  return SGVec3<float>(cart(0), cart(1), cart(2));
+  return SGVec3<float>((float)cart(0), (float)cart(1), (float)cart(2));
 }
 
 /// Unary +, do nothing ...
